requirements.txt
blinker==1.9.0
chardet==5.2.0
click==8.3.0
colorama==0.4.6
flask==3.1.2
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.3
python-dotenv==1.2.1
watchdog==6.0.0
werkzeug==3.1.3
